pyproject.toml
[project]
name = "stack-cli"
version = "0.1.0"
description = "AI-Human Engineering Stack — a layered cognitive pipeline for Claude Code CLI"
requires-python = ">=3.12"
dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
stack = "app.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]
artifacts = ["src/app/harness/**"]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]
